chore: BlockInput을 BlockTitleInput으로 변경 및 container로 이동(#23)
